test/system: pasta_test_do add explicit port check
Do not rely on an arbitrary delat in order to ensure the port was bound
in the container. Insted this approach checks if the port is bound in
the netns and only then starts the client. This speeds up the entire
test file by 50% but more importanly in parallel testing it solves hangs
as the timeout there was unreliable.

Fixes #23471

Signed-off-by: Paul Holzinger <pholzing@redhat.com>