podman artifact
the podman artifact verb is used to manage OCi artifacts. inspect is the
first of several sub-verb commands we need.  this pr introduces the
command and the neccesary plumbing to do the rest of the artifact
commands without an overwhelmingly huge PR.

until more sub commands are implemented, e2e tests are not possible.

there are TODOs in this code; most of which cannot be firmly established
at the moment and therefore are as such.

cmd/podman/artifact: add shell completion funcs

Also fix up some Args function to specify the correct number of args.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
Signed-off-by: Brent Baude <bbaude@redhat.com>